requirements.txt
flask
ultralytics
werkzeug
Pillow